// src/cloud/components/molecules/Editor/types.ts
export type FormattingTool =
  | 'header1'
  | 'header2'
  | 'header3'
  | 'header4'
  | 'header5'
  | 'header6'
  | 'admonitionNote'
  | 'admonitionInfo'
  | 'admonitionTip'
  | 'admonitionImportant'
  | 'admonitionDanger'
  | 'admonitionWarning'
  | 'admonitionCaution'
  | 'admonitionSecondary'
  | 'bold'
  | 'italic'
  | 'quote'
  | 'code'
  | 'bulletedList'
  | 'numberedList'
  | 'taskList'
  | 'link'
  | 'codefence'